introduce dropdown to typescript
