+++ src/components/Dropdowns/UserDropdown.tsx
@@ -1,14 +1,15 @@
-import React from "react";
+import React, { RefObject } from "react";
 import { createPopper } from "@popperjs/core";
 
-const NotificationDropdown = () => {
+const UserDropdown = () => {
   // dropdown props
   const [dropdownPopoverShow, setDropdownPopoverShow] = React.useState(false);
-  const btnDropdownRef = React.createRef();
-  const popoverDropdownRef = React.createRef();
+  const btnDropdownRef = React.createRef() as RefObject<HTMLAnchorElement>;
+  const popoverDropdownRef = React.createRef() as RefObject<HTMLDivElement>;
   const openDropdownPopover = () => {
+    if (btnDropdownRef.current && popoverDropdownRef.current)
     createPopper(btnDropdownRef.current, popoverDropdownRef.current, {
-      placement: "left-start",
+      placement: "bottom-start",
     });
     setDropdownPopoverShow(true);
   };
@@ -18,7 +19,7 @@ const NotificationDropdown = () => {
   return (
     <>
       <a
-        className="text-blueGray-500 py-1 px-3"
+        className="text-blueGray-500 block"
         href="#pablo"
         ref={btnDropdownRef}
         onClick={(e) => {
@@ -26,7 +27,15 @@ const NotificationDropdown = () => {
           dropdownPopoverShow ? closeDropdownPopover() : openDropdownPopover();
         }}
       >
-        <i className="fas fa-ellipsis-v"></i>
+        <div className="items-center flex">
+          <span className="w-12 h-12 text-sm text-white bg-blueGray-200 inline-flex items-center justify-center rounded-full">
+            <img
+              alt="..."
+              className="w-full rounded-full align-middle border-none shadow-lg"
+              src={require("assets/img/team-1-800x800.jpg").default}
+            />
+          </span>
+        </div>
       </a>
       <div
         ref={popoverDropdownRef}
@@ -62,9 +71,19 @@ const NotificationDropdown = () => {
         >
           Something else here
         </a>
+        <div className="h-0 my-2 border border-solid border-blueGray-100" />
+        <a
+          href="#pablo"
+          className={
+            "text-sm py-2 px-4 font-normal block w-full whitespace-nowrap bg-transparent text-blueGray-700"
+          }
+          onClick={(e) => e.preventDefault()}
+        >
+          Seprated link
+        </a>
       </div>
     </>
   );
 };
 
-export default NotificationDropdown;
+export default UserDropdown;
